# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gerenciamento de fazendas leiteiras\nSistema completo para gestão de animais, coletas de leite, reproduções, vendas e dívidas",
        "title": "FazendaPro API",
        "contact": {
            "name": "Suporte FazendaPro",
            "email": "suporte@fazendapro.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/farm": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados de uma fazenda específica pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm"
                ],
                "summary": "Obter fazenda por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da fazenda",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de uma fazenda existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "farm"
                ],
                "summary": "Atualizar fazenda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da fazenda",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados da fazenda",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateFarmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/debts": {
            "get": {
                "description": "Retorna lista paginada de dívidas com filtros opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Obter dívidas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número da página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Itens por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ano",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por mês",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DebtListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo registro de dívida",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Criar dívida",
                "parameters": [
                    {
                        "description": "Dados da dívida",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateDebtRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.DebtResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debts/{id}": {
            "delete": {
                "description": "Remove uma dívida do sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Deletar dívida",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da dívida",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreateDebtRequest": {
            "description": "Dados necessários para criar uma nova dívida",
            "type": "object",
            "properties": {
                "person": {
                    "description": "Nome da pessoa",
                    "type": "string",
                    "example": "João Silva"
                },
                "value": {
                    "description": "Valor da dívida",
                    "type": "number",
                    "example": 1500.5
                }
            }
        },
...
        "handlers.DebtListResponse": {
            "description": "Resposta com lista paginada de dívidas",
            "type": "object",
            "properties": {
                "debts": {
                    "description": "Lista de dívidas",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.DebtResponse"
                    }
                },
                "limit": {
                    "description": "Limite por página",
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "description": "Página atual",
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "description": "Total de registros",
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handlers.DebtResponse": {
            "description": "Resposta com dados de uma dívida",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Data de criação",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "id": {
                    "description": "ID da dívida",
                    "type": "integer",
                    "example": 1
                },
                "person": {
                    "description": "Nome da pessoa",
                    "type": "string",
                    "example": "João Silva"
                },
                "updated_at": {
                    "description": "Data de atualização",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "value": {
                    "description": "Valor da dívida",
                    "type": "number",
                    "example": 1500.5
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.UpdateFarmRequest": {
            "description": "Dados para atualizar uma fazenda",
            "type": "object",
            "properties": {
                "logo": {
                    "description": "Logo da fazenda em base64",
                    "type": "string",
                    "example": "data:image/png;base64,..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e depois o token JWT. Exemplo: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}